nexus-cdc/pyproject.toml
[project]
name = "nexus-cdc"
version = "0.1.0"
description = "Nexus MongoDB change stream listener — Debezium-lite emits CDC events to S3 for Databricks Autoloader"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.7",
    "boto3>=1.34",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "python-ulid>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "mongomock>=4.2",
    "moto[s3,dynamodb]>=5.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_cdc"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
